+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.5.0 (unreleased)
+## 1.0.0
 
 ### Added